dev: write migrations

+++ internal/models/api.go
@@ -8,7 +8,7 @@ type AuthenticationDataRequest struct {
 type OrderResponse struct {
 	Number     string      `json:"number"`
 	Status     OrderStatus `json:"status"`
-	Accrual    int         `json:"accural"`
+	Accrual    float64     `json:"accural"`
 	UploadedAt string      `json:"uploaded_at"`
 }
